requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlmodel>=0.0.20
python-dotenv>=1.0.0
PyJWT>=2.0.0
passlib>=1.7.0
bcrypt>=4.0.0
twilio>=8.0.0
httpx>=0.25.0
requests>=2.30.0
python-multipart>=0.0.6
jinja2>=3.1.0
aiofiles>=23.0.0
websockets>=10.0
redis>=4.0.0
pytz>=2022.1
sqlalchemy>=2.0.0
pydantic>=1.10.0,<3.0.0
packaging>=22.0
anthropic>=0.18.0